+++ drivers/infiniband/hw/qib/qib_qp.c
@@ -102,7 +102,7 @@ static u32 credit_table[31] = {
 
 static void get_map_page(struct qib_qpn_table *qpt, struct qpn_map *map)
 {
-	unsigned long page = get_zeroed_page(GFP_KERNEL);
+	unsigned long page = (unsigned long)get_zeroed_page(GFP_KERNEL);
 
 	/*
 	 * Free the page if someone raced with us installing it.
