+++ drivers/lguest/page_tables.c
@@ -319,7 +319,7 @@ static pte_t *find_spte(struct lg_cpu *cpu, unsigned long vaddr, bool allocate,
 		if (!allocate)
 			return NULL;
 
-		ptepage = get_zeroed_page(GFP_KERNEL);
+		ptepage = (unsigned long)get_zeroed_page(GFP_KERNEL);
 		/*
 		 * This is not really the Guest's fault, but killing it is
 		 * simple for this corner case.
@@ -351,7 +351,7 @@ static pte_t *find_spte(struct lg_cpu *cpu, unsigned long vaddr, bool allocate,
 		if (!allocate)
 			return NULL;
 
-		ptepage = get_zeroed_page(GFP_KERNEL);
+		ptepage = (unsigned long)get_zeroed_page(GFP_KERNEL);
 
 		/*
 		 * This is not really the Guest's fault, but killing it is
